requirement.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
django-ckeditor>=6.7.0
gunicorn>=21.2.0
redis>=5.0.1  # If using Redis for caching
django-environ>=0.11.2  # For managing environment variables
googletrans==4.0.0-rc1  # For translation support
pytest>=8.3.4  # For running tests
pytest-django>=4.8.0  # Django-specific pytest support
flake8>=7.0.0  # For linting and code style checking
